feat: refactor Set and SetCollection for improved centroid calculations and add new taxonomization features
